refactor: applicant 패키지 형식 수정
