+++ src/main/java/com/unionmate/backend/domain/applicant/application/dto/request/AnswerRequest.java
@@ -1,4 +1,4 @@
-package com.unionmate.backend.domain.applicant.dto.request;
+package com.unionmate.backend.domain.applicant.application.dto.request;
 
 import jakarta.validation.constraints.NotNull;
