+++ src/main/java/com/unionmate/backend/domain/applicant/application/dto/request/CreateApplicantRequest.java
@@ -1,4 +1,4 @@
-package com.unionmate.backend.domain.applicant.dto.request;
+package com.unionmate.backend.domain.applicant.application.dto.request;
 
 import java.util.List;
