+++ src/main/java/com/unionmate/backend/domain/applicant/domain/entity/column/Answer.java
@@ -1,0 +1,7 @@
+package com.unionmate.backend.domain.applicant.domain.entity.column;
+
+public record Answer<T>(
+    T answer
+) {
+
+}
