+++ src/main/java/com/unionmate/backend/domain/applicant/domain/entity/enums/EvaluationStatus.java
@@ -1,0 +1,7 @@
+package com.unionmate.backend.domain.applicant.domain.entity.enums;
+
+public enum EvaluationStatus {
+  SUBMITTED,
+  FAILED,
+  PASSED
+}
