--- src/main/java/com/unionmate/backend/domain/applicant/entity/column/Answer.java
@@ -1,7 +1,0 @@
-package com.unionmate.backend.domain.applicant.entity.column;
-
-public record Answer<T>(
-    T answer
-) {
-
-}
