--- src/main/java/com/unionmate/backend/domain/applicant/entity/enums/EvaluationStatus.java
@@ -1,7 +1,0 @@
-package com.unionmate.backend.domain.applicant.entity.enums;
-
-public enum EvaluationStatus {
-  SUBMITTED,
-  FAILED,
-  PASSED
-}
