+++ src/main/java/com/unionmate/backend/global/converter/LocalDateAnswerConverter.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.unionmate.backend.domain.applicant.entity.column.Answer;
+import com.unionmate.backend.domain.applicant.domain.entity.column.Answer;
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 import java.time.LocalDate;
